Break tx components into signle purposed module files
