+++ src/modules/transaction/components/TransactionDetails/TransactionVotes.css
