+++ src/modules/transaction/components/TransactionDetails/amount.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import { getTxAmount } from '@transaction/utils/transaction';
+import DiscreetMode from 'src/modules/common/components/discreetMode';
+import TokenAmount from '@token/fungible/components/tokenAmount';
+import TransactionDetailsContext from '../../context/transactionDetailsContext';
+import ValueAndLabel from './valueAndLabel';
+import styles from './TransactionDetails.css';
+
+const Amount = ({ t }) => {
+  const { activeToken, transaction } = React.useContext(
+    TransactionDetailsContext,
+  );
+  const addresses = [
+    transaction.asset.recipient?.address ?? '',
+    transaction.sender.address,
+  ];
+
+  return (
+    <ValueAndLabel label={t('Amount of transaction')} className={styles.amount}>
+      <DiscreetMode addresses={addresses} shouldEvaluateForOtherAccounts>
+        <span className="tx-amount">
+          <TokenAmount val={getTxAmount(transaction)} />
+          {' '}
+          {activeToken}
+        </span>
+      </DiscreetMode>
+    </ValueAndLabel>
+  );
+};
+
+export default Amount;
