+++ src/modules/transaction/components/TransactionDetails/index.js
@@ -1,2 +1,17 @@
-export * from './TransactionDetailComponents';
-export { default as TransactionVotes } from '../TransactionVotes';
+export { default as Amount } from './amount';
+export { default as BlockHeight } from './blockHeight';
+export { default as BlockId } from './blockId';
+export { default as Confirmations } from './confirmations';
+export { default as Fee } from './fee';
+export { default as Illustration } from './illustration';
+export { default as Members } from './members';
+export { default as Message } from './message';
+export { default as Nonce } from './nonce';
+export { default as PrettyJson } from './prettyJson';
+export { default as Recipient } from './recipient';
+export { default as RequiredSignatures } from './requiredSignatures';
+export { default as Sender } from './sender';
+export { default as SignedAndRemainingMembers } from './signedAndRemainingMembers';
+export { default as TransactionId } from './transactionId';
+export { default as TxDate } from './txDate';
+export { Votes } from './votes';
