+++ src/modules/transaction/components/TransactionDetails/message.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { getTxAsset } from '@transaction/utils';
+import TransactionDetailsContext from '../../context/transactionDetailsContext';
+import ValueAndLabel from './valueAndLabel';
+import styles from './TransactionDetails.css';
+
+const Message = ({ t }) => {
+  const { transaction } = React.useContext(TransactionDetailsContext);
+
+  return (
+    <ValueAndLabel label={t('Message')} className={styles.message}>
+      <div className="tx-reference">{getTxAsset(transaction)}</div>
+    </ValueAndLabel>
+  );
+};
+
+export default Message;
