+++ src/modules/transaction/components/TransactionDetails/votes.js
@@ -5,7 +5,7 @@ import TransactionDetailsContext from '../../context/transactionDetailsContext';
 import styles from './TransactionVotes.css';
 import VoteItem from '../VoteItem';
 
-export const TransactionVotesComp = ({ t, votedDelegates }) => {
+export const Votes = ({ t, votedDelegates }) => {
   const { transaction } = React.useContext(TransactionDetailsContext);
   const { votes } = transaction.asset;
 
@@ -54,4 +54,4 @@ export default withData({
       return responseMap;
     },
   },
-})(TransactionVotesComp);
+})(Votes);
