+++ src/modules/transaction/components/TransactionDetails/votes.test.js
