--- src/modules/transaction/components/TransactionVotes/index.js
@@ -1,1 +1,0 @@
-export { default } from './TransactionVotes';
