+++ src/modules/transaction/components/transactionDetailsView/styles.css
@@ -1,0 +1,3 @@
+.container {
+  padding: 0 20px;
+}
